+++ app/src/main/AndroidManifest.xml
@@ -16,17 +16,17 @@
 <!--            android:label="@string/title_activity_repair"-->
 <!--            android:theme="@style/Theme.RepairTool.NoActionBar" />-->
         <activity
-            android:name=".repairman.RepairmanActivity"
+            android:name="com.kahuis.repairtool.presentation.repairman.RepairmanActivity"
             android:exported="true"
             android:label="RepairmanActivity"
             android:theme="@style/Theme.RepairTool.NoActionBar" />
         <activity
-            android:name=".repairman.EditRepair"
+            android:name="com.kahuis.repairtool.presentation.repairman.EditRepair"
             android:exported="true"
             android:label="EditRepair"
             android:theme="@style/Theme.RepairTool.NoActionBar"/>
         <activity
-            android:name=".MainActivity"
+            android:name="com.kahuis.repairtool.presentation.MainActivity"
             android:exported="true"
             android:theme="@style/Theme.RepairTool.NoActionBar">
             <intent-filter>
