--- app/src/main/java/com/example/repairtool/repairs/Repair.kt
@@ -1,10 +1,0 @@
-package com.example.repairtool.repairs
-
-data class Repair(
-    val id: Int,
-    val name: String,
-    var building: String,
-    var priority: String,
-    var description: String,
-    var status: String
-    )
