--- app/src/main/java/com/example/repairtool/ui/theme/Color.kt
@@ -1,15 +1,0 @@
-package com.example.repairtool.ui.theme
-
-import androidx.compose.ui.graphics.Color
-
-//Light theme
-val CorporateRed = Color(0xFFff5757)
-val CorporateBlack = Color(0xFF3f3e3e)
-val CorporateWhite = Color(0xFFffffff)
-val CorporateGrey = Color(0xFFcecece)
-
-//Dark theme
-val CorporateRedDark = Color(0xFFD62F2F)
-val CorporateGreyDark = Color(0xFF6E6E6E)
-val CorporateBlackDark = Color(0xFF000000)
-
