--- app/src/main/java/com/example/repairtool/utilities/dropdownmenu/items/Building.kt
@@ -1,5 +1,0 @@
-package com.example.repairtool.utilities.dropdownmenu.items
-
-data class Building(val name:String) {
-    override fun toString() = name
-}
