--- app/src/main/java/com/example/repairtool/utilities/dropdownmenu/items/BuildingList.kt
@@ -1,12 +1,0 @@
-package com.example.repairtool.utilities.dropdownmenu.items
-
-object BuildingList {
-    var buildingList = mutableListOf(
-        Building(""),
-        Building("Damestoilet"),
-        Building("Hal"),
-        Building("Herentoilet"),
-        Building("Hoofdgebouw"),
-        Building("Kantoor")
-    )
-}
