--- app/src/main/java/com/example/repairtool/utilities/dropdownmenu/items/Priority.kt
@@ -1,5 +1,0 @@
-package com.example.repairtool.utilities.dropdownmenu.items
-
-data class Priority(val priority: String) {
-    override fun toString() = priority
-}
