--- app/src/main/java/com/example/repairtool/utilities/dropdownmenu/items/PriorityList.kt
@@ -1,10 +1,0 @@
-package com.example.repairtool.utilities.dropdownmenu.items
-
-object PriorityList {
-    var priorityList = mutableListOf(
-        Priority(""),
-        Priority("Hoog"),
-        Priority("Middel"),
-        Priority("Laag")
-    )
-}
