--- app/src/main/java/com/example/repairtool/utilities/dropdownmenu/items/Status.kt
@@ -1,5 +1,0 @@
-package com.example.repairtool.utilities.dropdownmenu.items
-
-data class Status(val status:String) {
-    override fun toString() = status
-}
