--- app/src/main/java/com/example/repairtool/utilities/dropdownmenu/items/StatusList.kt
@@ -1,11 +1,0 @@
-package com.example.repairtool.utilities.dropdownmenu.items
-
-object StatusList {
-    var statusList = mutableListOf(
-        Status(""),
-        Status("Geregistreerd"),
-        Status("In behandeling"),
-        Status("Wachtend"),
-        Status("Afgehandeld"),
-    )
-}
