+++ app/src/main/java/com/kahuis/repairtool/common/Constants.kt
@@ -1,0 +1,5 @@
+package com.kahuis.repairtool.common
+
+object Constants {
+    val urlApi = "http://yoururlhere.com"
+}
