+++ app/src/main/java/com/kahuis/repairtool/common/utilities/dropdownmenu/items/Building.kt
@@ -1,0 +1,5 @@
+package com.kahuis.repairtool.common.utilities.dropdownmenu.items
+
+data class Building(val name:String) {
+    override fun toString() = name
+}
