+++ app/src/main/java/com/kahuis/repairtool/common/utilities/dropdownmenu/items/BuildingList.kt
@@ -1,0 +1,12 @@
+package com.kahuis.repairtool.common.utilities.dropdownmenu.items
+
+object BuildingList {
+    var buildingList = mutableListOf(
+        Building(""),
+        Building("Damestoilet"),
+        Building("Hal"),
+        Building("Herentoilet"),
+        Building("Hoofdgebouw"),
+        Building("Kantoor")
+    )
+}
