+++ app/src/main/java/com/kahuis/repairtool/common/utilities/dropdownmenu/items/Priority.kt
@@ -1,0 +1,5 @@
+package com.kahuis.repairtool.common.utilities.dropdownmenu.items
+
+data class Priority(val priority: String) {
+    override fun toString() = priority
+}
