+++ app/src/main/java/com/kahuis/repairtool/common/utilities/dropdownmenu/items/PriorityList.kt
@@ -1,0 +1,10 @@
+package com.kahuis.repairtool.common.utilities.dropdownmenu.items
+
+object PriorityList {
+    var priorityList = mutableListOf(
+        Priority(""),
+        Priority("Hoog"),
+        Priority("Middel"),
+        Priority("Laag")
+    )
+}
