+++ app/src/main/java/com/kahuis/repairtool/common/utilities/dropdownmenu/items/Status.kt
@@ -1,0 +1,5 @@
+package com.kahuis.repairtool.common.utilities.dropdownmenu.items
+
+data class Status(val status:String) {
+    override fun toString() = status
+}
