+++ app/src/main/java/com/kahuis/repairtool/common/utilities/dropdownmenu/items/StatusList.kt
@@ -1,0 +1,11 @@
+package com.kahuis.repairtool.common.utilities.dropdownmenu.items
+
+object StatusList {
+    var statusList = mutableListOf(
+        Status(""),
+        Status("Geregistreerd"),
+        Status("In behandeling"),
+        Status("Wachtend"),
+        Status("Afgehandeld"),
+    )
+}
