+++ app/src/main/java/com/kahuis/repairtool/presentation/repairs/RepairList.kt
@@ -1,0 +1,9 @@
+package com.kahuis.repairtool.presentation.repairs
+
+object RepairList {
+    var repairList = mutableListOf(
+        Repair(1,"Lamp stuk", "Hoofdgebouw","Hoog", "Dit is een hardcodedtest", "Geregistreerd"),
+        Repair(2, "Wc spoelt niet door", "Herentoilet", "Middel", "Dit is een hardcodedtest", "In behandeling"),
+        Repair(3, "Deur klemt", "Kantoor","Laag", "Dit is een hardcodedtest", "Wachtend")
+    )
+}
